Remove WithAbortAsync
It's a hold-over from before I was aware that canceling an individual operation should actually abort the whole websocket.  Now that that's the case, it's much cheaper to just register with the token as part of the async operations, rather than wrapping each in another task that itself does such registration.