Use custom network stream in managed WebSockets on Unix
We only need to support a single read and a single write at a time, which means we can create two SocketAsyncEventArgs instances and repeatedly reuse them for all read and write operations on the stream.  This avoids a significant amount of the overhead involved in reading/writing with the network.