Optimize SendFrameAsync to minimize allocations
The previous commit was beneficial in avoiding one large set of overhead, but it also has a flaw that, due to a current behavior in Sockets, sending to a socket inevitably completes asynchronously, which means SendFrameAsync almost always hits an await that yields.  Since this is a very important code path, until the Socket issue can be addressed, we can optimize the common case to minimize the costs involved.