Reduce System.Threading.Tasks.Tests from ~140s to ~10s
As with https://github.com/dotnet/corefx/pull/9468 and https://github.com/dotnet/corefx/pull/9466, this commit focuses on reducing the total, outerloop running time of our worst-offending test suites.  This one is on the System.Threading.Tasks tests.  Some of the tests are not in good shape, e.g. it's not clear to me that the TaskCancelWaitTests are actually verifying everything they should, even though they're spending a lot of time doing it.

Changes:
- Reduced artificially long wait times in TaskAwaiter, IAsyncResult, TaskCancelWaitTests, TaskStatusTests, CESchedulerPairTests, and Coverage tests
- Reduced size of workload in TaskCancelWaitTests, TaskContinueWithAllAnyTests, and TaskWaitAllAnyTests
- Ensured that the ThreadPool was ramped up to serve several tests expecting more threads available
- Removed some unnecessary spinning