Make X509 CollectionsTests run without modifying system state
X509ChainElementCollection_CopyTo_NonZeroLowerBound_ThrowsIndexOutOfRangeException naively builds a cert chain, but since the issuer
isn't part of the ca-certificates bundle it gets downloaded as part of the
test (and then cached to the hard drive).

Since the network-enabled and caching parts are not explicitly needed for this test, provide it enough information to work fully-standalone.