Add ReferenceEqualityComparer to Common
In X509Chain building some internally-managed X509Certificate2 objects
are mixed in with some user-owned objects. After mining for the required
data any unused internally-managed objects should be Disposed, but it's
not straightforward to identify if it's a internal object or a user object being
passed to the next stage of the pipeline because they .Equals the same.