Reduce System.Runtime.Numerics outerloop test time from ~180s to ~90s
There are two remaining individual tests each taking ~80s doing lots of calculations (RunDoubleExplicitCastFromLargeBigIntegerTests and RunLargeValueLogTests)... I don't know the importance enough of what those are testing to determine whether it's ok to change them, so I've left them as is, which means it's going to be difficult to significantly reduce the overall running time (which on my machine is now ~90s) much further.  Still, this small change brings the running time on my machine down from ~150 seconds to ~90 seconds.

Changes:
- The ModPow3LargeInt test was generated 10 sets of random values, using BigInteger.ModPow with them, and then comparing the result against another, very inefficient implementation of modpow, resulting in a test taking ~150 seconds (and another one calling it with slightly different inputs taking ~30s).  I simply pre-generated a set of 10 inputs and expected outputs (using the same algorithm the test was previously using), and put those into the test as InlineData.