Reduce S.R.InteropServices outerloop test time from ~100s to ~0.2s
Most of the tests in the test library run very quickly.  But one runs for ~100s, calling HandleCollector.Add 2B times to get it to overflow.  As a hack but IMHO a worthwhile hack, this commit uses reflection to jump the HandleCollector forward as if it were already called almost int.MaxValue times.  It then proceeds to do the remaining Adds manually, verifying that still works, and then overflows and verifies the expected failure.  This reduces the overall test execution from over 100 seconds to a few hundred milliseconds.