Convert CredentialCache perf tests to xunit.performance
These tests were running for a long time, and then ignoring the result.  These tests run for ~2 minutes in our test passes, even though we're not getting useful data out of them.

This change converts the tests to use xunit.performance.  This will greatly reduce their runtime when not running as performance tests (we just run one iteration in that case).  It will also enable them to produce useful results in our perf lab, where we *do* have the infrastructure to find regressions in both throughput and GC counts for xunit.performance-based tests.