Removed redundant calls to SetFieldsFromUri() in UriBuilder
SetFieldsFromUri() is more or less identical to Init() (which for some reason calls SetFieldsFromUri() resulting in UriBuilder being initialized twice). The only other place it's called is in the Uri property after the new Uri object has been constructed. It doesn't make sense for a UriBuilder to re-initialize itself with the information it just used to construct a new Uri object.