Move Dataflow ETW test into its own process
Fixes an issue where if another test accidentally overlaps (e.g. if the test doesn't ensure all queued async operations have completed), ETW events may fire in a way not expected by the tests.

This also eliminates the reason we had disabled parallelism for the suite, so I re-enabled it, which also results in the tests running several seconds faster.