Fix logic with setting response body receive timeout
Due to code moving around with PR #8996, the `state.RequestHandle` was
already null when trying to set the receive timeout for the response
body. This generated an exception which was caught internally.

This fix moves the setting of the receive timeout so that it has a valid
`state.RequestHandle` to use.

Fixes #9578