Add Retain to OSX FileSystemWatcher runloop
The OSX FileSystemWatcher uses the CoreFoundation RunLoop API to watch for events. There is a rule that CF objects recieved from getter functions are not guaranteed to exist unless explicitly retained. We were not doing this, so there was a very infrequent potential race condition wherein cancellation raced with completion of the thread that owned the runloop. When completion happened first, OSX disposed the unmanaged CoreFoundation RunLoop object that was unique to the thread, and cancellation attempted to call API on that now-invalid object. This commit adds explicit retain/release logic to prevent that race condition.