Fix Console.In.Peek and friends
The StdInStreamReader used on Unix as the TextReader used in Console.In derives from StreamReader and is passed the ConsoleStream for stdin.  This is both unnecessary and problematic.  It's unnecessary because all of the reading is done via special functions exposed from System.Native.so.  It's problematic because the base StreamReader does its own buffering based on the ConsoleStream and not all of the functions have been overridden by the StdInStreamReader.  As a result, for example, calling Read(char[], int, int) ends up bypassing previously cached data.

This commit does two main things:
- Renames StdInStreamReader to StdInReader and changes it from deriving from StreamReader to deriving from the base TextReader.  The only functions that must be overridden on TextReader to enable the behaviorally enable the rest are Read and Peek.
- Adds an override of Peek.  Read is already overridden.