Only build a subset of packages
In this release branch we only want to build that packages that we
intend to ship.

Set pre-release to servicing so that packages that are not stable
get a sensible pre-release suffix.

Turn off package version baseline consistency check.  We want our
baseline to be frozen at RTM (otherwise every package built is a
patch family dependency).