Remove some random data from System.Reflection.Emit.Tests
Currently, we generate random data in some of the System.Reflection.Emit
tests
One example is om the test for PropertyBuilder.Name that generates a
random char.

Due to bug #7166 xunit crashes when there is an invalid char (e.g. lone
high surrogate) in a MemberData or InlineData

Should hopefully fix #9820