Add a test for Environment.ExitCode
We've consolidated almost all of our remote execution logic (testing code in another process) into a single shared console app that supports using reflection to invoke delegates, but to test ExitCode properly, we need t a void-returning Main method.  This commit adds such a little helper app to the System.Runtime.Extensions tests suite and adds a test that uses it.  That test is disabled until the associated issue in coreclr is fixed.