Add ICloneable back to System.Runtime contract
We've got a bunch of types outside of mscorlib that implement this interface in desktop, and the lack of it is causing issues, e.g.  https://github.com/dotnet/corefx/issues/9884.  This commit adds the interface to the contract; when we subsequently update the packages, then we can start implementing this in types outside of System.Runtime (since System.Runtime is mostly a facade on System.Private.Corelib, the types it "contains" already implement the interface).