Fix FileStream.Flush on Unix when wrapping non-synchronizable descriptors
FileStream can be used with any arbitrary file descriptor on Unix.  FileStream.Flush may use FSync to try to synchronize the file's state with the storage device, but that only makes sense and works for certain kinds of file descriptors.  For other file descriptors, currently we're throwing an exception if the operation fails, when really we should just ignore the failure, as there's no work to be done.