MEF2 throws a System.TypeLoadException
Make sure the typeArguments follow the generic type parameters
constraints.
If not, then the part is not a match and must be ignored.
Otherwise, _partType.MakeGenericType throws a System.TypeLoadException.
(Unlike MEF2, MEF on the full .NET framework filters correctly and
ignores parts that do not match the constraints, so this fix makes the
behavior the same in both versions).