Remove two Sockets tests
These tests try to connect to an endpoint that is not supposed to have an active listener, and expect to fail.  However, there's no guarantee that some other test (or other process) is not listening on that endpoint.  This leads to occasional failures in these tests (and may trigger failure in the *other* test that happens to have been allocated the port in question).

I don't know of a way to reliably force the failure these test cases are looking for, and I don't believe these tests provide enough value to justify "heroic" effor here.  So I'm simply removing them.