Turned tests that use KeyAgreement certs or add OriginatorCerts into ConditionalFacts in EnvelopedCms
As KeyAgreement certs can't be used for encryption or decryption on the
current implementation of EnvelopedCms for Unix, the tests were turned
into conditional facts and tests were added to ensure that the Unix
implementation throws a PlatformNotSupportedException.

Similarly tests that added originator certificates were made
ConditionalFacts as OpenSSL doesn't support adding these and a test was
added to ensure that the Unix implementation throws a
PlatformNotSupportedException.