Remove TODO comment from Socket ArgumentValidation tests
The Accept/Connect suggestions in this comment don't make much sense in the current API, as there is no "disconnect" API in CoreFx.

The End* suggestions would be great, except that the public Begin*/End* methods are implemented using generic wrappers over Task-based async methods, and so are not likely to ever pass such tests.  It doesn't seem worth complicating the implementation for this.