Implement Environment.Get/SetEnvironmentVariable(s) for User and Machine on Win32
The implementations are currently commented out because they rely on the registry.  It looks like we're going to land in a place where Environment doesn't have access to Microsoft.Win32.Registry, so we need an alternative.  Various options include:
1. Exposing the Registry implementation from coreclr for System.Runtime.Extensions to use
2. Using Microsoft.Win32.Registry.dll via reflection.
3. Writing a custom implementation of registry access for System.Runtime.Extensions.
4. Compiling the relevant code from Microsoft.Win32.Registry into System.Runtime.Extensions, with the types as internal in S.R.E and public in M.W.R.

I've gone with (4).  It's all implementation detail, so we can change it in the future if an alternative approach proves better.