ARM-CI : Check a device is already mounted while mounting a device.
Now ARM-CI makes building failure frequently with below messages.

http://dotnet-ci.cloudapp.net/job/dotnet_corefx/job/master/job/linuxarmemulator_cross_release/127/console
19:21:33 + unmount_rootfs /opt/linux-arm-emulator-root/dev
19:21:33 + local rootfsFolder=/opt/linux-arm-emulator-root/dev
19:21:33 + grep -qs /opt/linux-arm-emulator-root/dev /proc/mounts
19:21:33 + sudo umount /opt/linux-arm-emulator-root/dev
umount: /opt/linux-arm-emulator-root/dev: device is busy.
19:21:33         (In some cases useful info about processes that use
19:21:33          the device is found by lsof(8) or fuser(1))
19:21:33 Build step 'Execute shell' marked build as failure

I suggest though the script is exited by any cases, the script would not run un-mounting.
But whenever CI is running and mounting a device, It will check the device is already mounted.

I think this error is occurred from the script of running tests. I hope to resolve this problem As soon as possible.
In before resolving this problem, It would rather apply it than errors are occurred.