Prepare for picking up package harvesting tools
https://github.com/dotnet/buildtools/pull/932

This ensures that we the last previously version of packages.

For System.Data.Common we needed to suppress netstandard inference since
that previous package supported wp8 and netcore45 but specifically could
not support net45 (and thus netstandard1.0 & 1.1) due to missing members
on the desktop inbox types.

For System.Security.Cryptography.Pkcs that package incorrectly listed
xamarin frameworks as supporting it inbox in its first release.  We're
intentionally ommitting those.