Update SNI package used for testing SqlClient
The sqlclient src package uses the last stable version of the runtime sni package, so I updated the hard-coded SNI versions used in testing to be the latest prerelease version. This way we're actually testing the latest SNI version that we're publishing. I also updated the ExternalPackages props to be correct for clrcompression and updated them for the new SNI package.