Move System.Diagnostics.Debug.SymbolReader to System.Diagnostics.StackTrace.
Moves the portable PDB support for SOS to the System.Diagnostics.StackTrace assembly to
be along side the mscorlib StackTraceHelper wrapper code.

Removed System.Diagnostics.Debug.SymbolReader.

Uses a GCHandle to pass the StackTraceSymbols/PDB reader to the various functions.

Added SOS in-memory PDB support.

Most of the plumbing for in-memory PE support is in place.