Add fastpath for restore
This picks up the latest buildtools which includes a task that adds a
fast path for restore.  We will now call that task to do a check up
front to determine if a restore is actually needed before invoking
restore.  This helps the common case where everything is up-to-date
reducing the restore from 5 minutes down to 1 second.