Fix Stream.Read/WriteAsync calls to FromAsync
The task created from FromAsync is a promise, such that it doesn't actually run any code.  As a result, it's invalid to supply DenyChildAttach or HideScheduler, as these are only relevant to tasks that run code (since they affect ambient state around that running code).  We should just use the default options.