RegistryKey.OpenSubKey should remove multiple/trailing slashes
After introducing the mini-PAL, the `RegistryKey.OpenSubKey(string)`
and `OpenSubKey(string, bool)` methods do not fix-up names that contain
multiple/trailing slashes, which is a behavior change.

Notably, `OpenSubKey(string, RegistryRights)` did not change, it
behaves as it should.

This commit fixes the issue, and adds more test coverage for the
following methods:

- `OpenSubKey`
- `CreateSubKey`
- `DeleteSubKey`
- `DeleteSubKeyTree`

The relevant new `OpenSubKey` tests fail before this change and pass
after. The new tests also pass when run on .NET Framework 4.6.2.