Implement digest and HMAC on CommonCrypto
Digest algorithms: MD5, SHA1, SHA-2-256, SHA-2-384, SHA-2-512
HMAC algorithms: MD5, SHA1, SHA-2-256, SHA-2-384, SHA-2-512

The rough design:
* Use a single PAL identifier for the algorithm across digest and HMAC.
* The same identifier will be used in RSA
* Apple's hashing primitives require 3 exports per algorithm.
Instead of exposing them directly, make a shim-internal struct.
* Apple's HMAC primitives require re-asserting the algorithm ID on reset,
instead of persisting it in a managed field, just carry it in a shim-internal struct.

Regarding the identifiers.  While HMAC has an enum for algorithm ID,
and we could use the HMAC algorithm ID as our hash algorithm ID, the
identifiers for RSA signatures are different (for one, they're strings instead
of numbers).  The algorithm identifiers for RSA are further different in that
they use a single "SHA2" value for algorithm, then disambiguate
256/384/512 based on a second value (hash input size). That, combined
with many identifiers across their cryptography libraries being documented
as "may change in a future version" (to get people to use the constant
expression instead of just writing the literal value) resulted in identifiers
matching our (and Windows' and OpenSSL's) notion of an algorithm and
not trying to maintain semantic alignment for pass-through benefits.