On Linux, we were reporting the starting *stack* address for a thread via ProcessThread.StartAddress.  This property is supposed to return the *code* address of the start of the thread.  We have no way of getting that info on Linux, so we'll just report it as IntPtr.Zero, as we already do on OSX (and apparently sometimes on Windows).
This change also reworks the test for this property to be more permissive of address values.  We already accepted IntPtr.Zero on Windows/OSX, and now on Linux.  We previously failed on "negative" address values, but this doesn't make sense, as addresses are really unsigned.  We really just need to allow any value here.