Fix DefaultWebProxy_VerifyDefaults_Success test
This test had an ordering constraint which was not enforced, but the
test case was getting lucky with the default sort used by xunit.

That default sort is a comparison of SHA of the test assembly name, type
name and method.  I previously changed the test assembly version which
caused this SHA to change and distrupted the order.

Fix the test to not depend on execution order.