Don't dispose the system certs store before making a system trust decision
The X509 finalization cleanup (#9510) went a little further than required.
For non-root certificates the cleanup logic was fine, but for a root certificate
the following erroneous logic happened:

* Build trusted certs list
* FindIssuer(cert)
  * This cert is self-signed, it has no candidate parents
  * (candidates.Count == 0)
* For every trusted cert which wasn't already known to be a candidate, Dispose it.
  * This disposed all of the trusted certs.
* Now ask "is the root cert of the chain in the trusted certs set?".
  * The answer is no, since everything was disposed.

The caller, which received the candidates list and the trusted certs list, was
already disposing the trusted certs, so there's no reason to do that in the
inner method.  That realization then lets the cleanup logic be simplified to no
longer have an indexing dependency.

The new test which has the same root cert in both the user and machine root
store showed that there was a "didn't get added to the HashSet"-induced
finalization, and that has been cleaned up (and guarded against). The new
Dispose logic has been proven resilient to the system trust list being populated
in a different order than the FindIssuer presentation list.