DeflateStream: Allow inflation without input
Our implementation of DeflateStream only allows inflation when there is input to pass to the inflater. Because of how the stream was coded, this would raise the potential issue where we would be out of input but there could still be inflated bytes cached ready to read. This commit removes the checks for valid input and allows inflation regardless of input.
- This does have the downside of reducing efficiency in the scenario where we are actually out of bytes to read. We were previously operating under the incorrect assumption that no input implies no output, but that isn't strictly true.