Target compatible NuGet moniker
This updates two projects which were targeting NuGet monikers which were
incompatible with how they were being packaged.

This prepares for a buildtools change which will allow us to remove
PackageTargetFramework & validate against NuGetTargetMoniker.

System.Runtime.CompilerServices.Unsafe was not using NuGet but I
specified NuGetTargetMoniker to be consistent with how it was being
packaged.  A future change will remove PackageTargetFramework from this
project once we have a buildtools update.

System.Runtime.Extensions was building against netcoreapp1.0 but
packaging as netstandard1.5.  It didn't matter in this case since it was
only referencing the coreclr targeting pack, but we'd flag this during
validation.