Fix ProcessThread.TestStartTimeProperty flakiness
The test grabs the process threads list and then iterates through it accessing the StartTime property on each.  The determination of start time happens lazily when StartTime is accessed, not when the collection is accessed, so if the thread goes away between the time we get the list and the time we access StartTime, it fails with an exception.  This commit makes the test more lenient.