Clean up serialization source in corefx
The core primitives are now in coreclr and corert, so other than for building System.Runtime.Serialization.Primitives for netstandard1.3, they're no longer needed.  Move the relevant files into that library, and delete the rest, except for ones we want to be able to use internally in libraries that need attributes and to avoid our needing to use ifdefs.