Move some flaky FSW tests to Outerloop
All of the FSW tests are potentially subject to intermittent failures because of the inherint flakiness of the underlying OS file system watching APIs. Although we do some manual stuff to get around that flakiness (e.g. retries, long timeouts), some scenarios aren't reliable enough to make it through. This commit moves those to Outerloop so they don't get in the way of our gated check in tests.